+++ bin/introspection.sh
@@ -5,5 +5,4 @@ rm -rf ../src/Serwisant/SerwisantApi/Types/SchemaInternal
 rm -rf ../src/Serwisant/SerwisantApi/Types/SchemaPublic
 rm -rf ../src/Serwisant/SerwisantApi/Types/SchemaService
 
-export SERWISANT_HOST="http://127.0.0.1:3000"
 php introspection.php