+++ src/Serwisant/SerwisantApi/Types/SchemaPublic/PublicMutation.php
@@ -55,6 +55,20 @@ class PublicMutation extends Types\RootType
   }
 
   /**
+   * @param AnonymousApplicantInput $applicant
+   * @param RepairInput $repair
+   * @param RepairItemInput[] $additionalItems
+   * @param string[] $temporaryFiles
+   * @param string $device
+   * @param AddressInput $address
+   * @return RepairCreationResult
+   */
+  public function createRepair(AnonymousApplicantInput $applicant, RepairInput $repair, array $additionalItems = array(), array $temporaryFiles = array(), string $device = null, AddressInput $address = null, $vars = array())
+  {
+     return $this->inputArgs('createRepair', array_merge($vars, ['applicant' => $applicant, 'repair' => $repair, 'additionalItems' => $additionalItems, 'temporaryFiles' => $temporaryFiles, 'device' => $device, 'address' => $address]));
+  }
+
+  /**
    * @param FileInput $file
    * @return TemporaryFileCreationResult
    */
@@ -67,13 +81,13 @@ class PublicMutation extends Types\RootType
    * @param AnonymousApplicantInput $applicant
    * @param TicketInput $ticket
    * @param string[] $temporaryFiles
-   * @param string[] $devices
+   * @param string $device
    * @param AddressInput $address
    * @return TicketCreationResult
    */
-  public function createTicket(AnonymousApplicantInput $applicant, TicketInput $ticket, array $temporaryFiles = array(), array $devices = array(), AddressInput $address = null, $vars = array())
+  public function createTicket(AnonymousApplicantInput $applicant, TicketInput $ticket, array $temporaryFiles = array(), string $device = null, AddressInput $address = null, $vars = array())
   {
-     return $this->inputArgs('createTicket', array_merge($vars, ['applicant' => $applicant, 'ticket' => $ticket, 'temporaryFiles' => $temporaryFiles, 'devices' => $devices, 'address' => $address]));
+     return $this->inputArgs('createTicket', array_merge($vars, ['applicant' => $applicant, 'ticket' => $ticket, 'temporaryFiles' => $temporaryFiles, 'device' => $device, 'address' => $address]));
   }
 
   /**