+++ src/Serwisant/SerwisantApi/Types/SchemaService/RepairInput.php
@@ -51,21 +51,9 @@ class RepairInput extends Types\Type
 
   /**
    * @var string
-   * Address where service should pick up a repair item. Leave empty if `delivery` is `PERSONAL`
-  */
-  public $pickUpAddress;
-
-  /**
-   * @var string
    * Defines how repaired item should be returned to customer. Return address is specified in `returnAddress`
   */
   public $collection;
-
-  /**
-   * @var string
-   * Address where repaired item should be returned. Leave empty if `collection` type is `PERSONAL`
-  */
-  public $returnAddress;
 
   /**
    * @var string
@@ -76,6 +64,18 @@ class RepairInput extends Types\Type
    * @var CustomFieldValueInput[]
   */
   public $customFields = [];
+  /**
+   * @var string
+   * Address where service should pick up a repair item. Leave empty if `delivery` is `PERSONAL`
+  */
+  public $pickUpAddress;
+
+  /**
+   * @var string
+   * Address where repaired item should be returned. Leave empty if `collection` type is `PERSONAL`
+  */
+  public $returnAddress;
+
   /**
    * @var string
   */